core, util: conditional deps from features
